backend/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.2
Pillow==10.4.0
gunicorn==22.0.0
python-decouple==3.8
psycopg2-binary==2.9.9
